requirements.txt
pandas
aiohttp
numpy
python-dotenv